chore(sl-y666): move capitalize and normalizeName into @satsuma/core/string-utils
Creates satsuma-core/src/string-utils.ts as the single home for pure text
utilities with no parser dependencies. capitalize() was duplicated identically
in lint-engine.ts and validate.ts; normalizeName() lived only in satsuma-cli
with no reuse path for future tooling.

- Adds @satsuma/core exports: capitalize, normalizeName
- Adds ./string-utils subpath export to core package.json
- lint-engine.ts and validate.ts import capitalize from @satsuma/core
- normalize.ts re-exports normalizeName from @satsuma/core; flattenFields and
  matchFields remain CLI-local (field-structure logic, not string utilities)
- normalizeName tests moved to satsuma-core/test/string-utils.test.js;
  retired from CLI normalize.test.js to avoid duplication
