+++ .tickets/sl-y666.md
@@ -1,6 +1,6 @@
 ---
 id: sl-y666
-status: open
+status: in_progress
 deps: []
 links: []
 created: 2026-03-30T06:29:33Z
@@ -41,3 +41,10 @@ Update normalize.ts to import from @satsuma/core or remove it if fully subsumed.
 - Each test has a leading comment explaining the property under test and why it matters (e.g. "normalizeName must treat 'First Name', 'first_name', and 'firstname' as equivalent — this is the contract relied on by the field matching logic")
 - Any redundant capitalize tests already present in CLI test files are removed rather than duplicated
 
+
+## Notes
+
+**2026-03-30T06:49:58Z**
+
+Cause: capitalize() was duplicated in lint-engine.ts and validate.ts; normalizeName() lived only in satsuma-cli/normalize.ts with no path to reuse.
+Fix: created satsuma-core/src/string-utils.ts exporting capitalize and normalizeName; CLI files import from @satsuma/core; normalizeName tests moved to core, redundant CLI copies retired. (commit pending)
