+++ tooling/satsuma-cli/test/normalize.test.js
@@ -1,31 +1,18 @@
 /**
  * normalize.test.js — Unit tests for src/normalize.js
+ *
+ * Tests the CLI-level field-matching logic (matchFields).
+ * normalizeName is tested in satsuma-core/test/string-utils.test.js — its
+ * behaviour is not re-tested here to avoid duplication.
  */
 
 import assert from "node:assert/strict";
 import { describe, it } from "node:test";
-import { normalizeName, matchFields } from "#src/normalize.js";
-
-describe("normalizeName", () => {
-  it("lowercases and strips underscores", () => {
-    assert.equal(normalizeName("FIRST_NM"), "firstnm");
-  });
-
-  it("strips hyphens", () => {
-    assert.equal(normalizeName("first-name"), "firstname");
-  });
-
-  it("FirstName matches first_name after normalization", () => {
-    assert.equal(normalizeName("FirstName"), normalizeName("first_name"));
-  });
-
-  it("MailingCity does NOT match city", () => {
-    assert.notEqual(normalizeName("MailingCity"), normalizeName("city"));
-  });
-});
+import { matchFields } from "#src/normalize.js";
 
 describe("matchFields", () => {
   it("matches fields by normalized name", () => {
+    // Verifies the full matching pipeline: normalization + lookup + result shape.
     const src = [{ name: "FirstName", type: "VARCHAR" }];
     const tgt = [{ name: "first_name", type: "VARCHAR" }];
     const result = matchFields(src, tgt);
@@ -36,6 +23,7 @@ describe("matchFields", () => {
   });
 
   it("returns source-only and target-only correctly", () => {
+    // Ensures unmatched fields on each side are correctly categorised.
     const src = [
       { name: "id", type: "INT" },
       { name: "email", type: "VARCHAR" },
